IOA Un-tag VLAN Configuration
IOA Interface resource was using in-correct command for removing the un-tag VLAN configuration.
This error was not observed as same VLAN was used for PXE in a given setup again and again.
Now with FCoE we have to replace the un-tag VLAN from PXE to the default VLAN for FIP-Snooping, configuration was failing.