Support for PMUX and VLT support for M1000e IOA
Removed the configuration to enable the default mode on the switch. This was leading to an extra reboot and the switch was loosing access in case of static IP address configuration.
Removed the validation to block the uplink configuration for M1000e IOAs. If IOA is selected for uplink configuration then PMUX mode will be enabled on the switch
If VLT enabled is selected and VLT interfaces are mapped then switch will be moved to VLT mode